+++ src/grafana_core/app/features/dimensions/types.ts
@@ -86,3 +86,5 @@ export enum ResourcePickerSize {
     SMALL = 'small',
     NORMAL = 'normal',
 }
+
+export interface DirectionDimensionOptions {}