+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog Mapgl
 
 ## 2.0.2
-- Grafana 12.4.0 fix-ref: bump dependencies. **icon picker** drop-down for folders list is visible upon switching to 'Url' tab and back to 'Folders'. Final fix depends on fixing the breaking changes in @grafana/ui: 12.4.0
+- fixes the broken icon folder selection in **Grafana 12.4.0.** 
+- rm deprecated 'config.panels' usage
 - more robust svg load for arrows under stricter Grafana environments (**v11.2.0**)
+
  
 ## 2.0.1
 - arrows config for edges in 'routed' view mode